@@ -7,11 +7,11 @@
     "@testing-library/react": "^11.1.0",
     "@testing-library/user-event": "^12.1.10",
     "materialize-css": "^1.0.0-rc.2",
-    "react": "^17.0.1",
+    "react": "^18.2.0",
     "react-bootstrap-validation": "^0.1.11",
-    "react-dom": "^17.0.1",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.1",
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.14.1",
+    "react-scripts": "^5.0.1",
     "server": "file:..",
     "web-vitals": "^0.2.4"
   },
